Implement and use lognormal amount distribution.
